// movies-web/src/app/models/movie.model.ts
export interface Movie {
    title: string;
    releaseYear: string;
    locations: string;
    productionCompany: string;
    distributor: string;
    director: string;
    writer: string;
    actor1: string;
    actor2: string;
    actor3: string;
    longitude: string;
    latitude: string;
    analysisNeighborhood: string;
    supervisorDistrict: string;
}